Update patch set 1
Patch Set 1:

Nader going to update the spec files in networking-cisco, and can add the link here.

Ihar, we could split up the services (one runs on compute the other on controller) but given the way we are currently including python-networking-cisco in disk images, i'm not sure that adds anything other then a little more confusion (extra puppet logic, extra doc, explaining the diff between the RPM's, etc).  My personal pref is to keep a single RPM.  
That said this is your wheelhouse, so I'll defer to you.

Patch-set: 1