Update patch set 1
Patch Set 1:

Yes, neutron.egg-info was a typo.

 As for not needed, I'm less sure about that for a couple reasons.
1.) if/when patches are applied on a downstream repo pbr should re-generate the egg-info files after patches are applied
2.) _Ideally_ i'd like to keep things consistent with openstack-packges/neutron

One is based on quick reading of BZ / git comments, the other just a preference.

Both are up for debate, I started this review mainly to figure out if this was the correct way to do it or not.

Patch-set: 1
Label: Code-Review=0